[Format version 2] Add FileMetadata to FileHandle for modes
This adds a FileMetadata as a part of the FileHandle.

This does make me question if FileHandle is now appropriately named, but
logically, the file metadata is a part of the file state which should
be considered changed along with the other properties of the FileHandle,
so this seems like the correct place to put this information for now.